requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2022.5.18
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
django-on-heroku==1.1.2
djongo==1.3.6
dnspython==2.2.1
docopt==0.6.2
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
mccabe==0.6.1
pipreqs==0.4.11
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pymongo==3.12.3
pytz==2022.1
requests==2.27.1
sqlparse==0.2.4
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
yarg==0.1.9
django-material==1.10.0
django-viewflow==1.11.0